Improved command responses
The bot will now post the link to the newly created events in a given channel (currently staff only) and ping the @Events role. If the event creation fails, it will post the http response and ping @Owner. The bot responds to the command with a link to the post.